Begin a forked approach
So here's what I'm thinking this time:

A parent process will create a pipe before forking. The parent will
poll on the read end of the pipe.

The child process will open the serial port. It will read and buffer characters
as they come in. After each character, the buffer will be evaluated for
viability as a valid packet. If the buffer is not viable, it will be discarded.
When a complete packet is received, it will be converted into a packet struct
and written through the pipe.

When the parent receives the packet, it will act on it accordingly.